Changed how url is saved in database

+++ controller/video.js
@@ -17,13 +17,17 @@ const getAllImages = async (req, res) => {
 const getImageById = async (req, res) => {
     const imageId = req.params.id
 
-    const image = await ImageModel.findById(imageId)
-
-    if (!image) {
-        return res.status(404).json({ status: "error", message: "imageid not found" });
+    ImageModel.findById(imageId)
+        .then((data) => {
+            if (!data) {
+                throw new Error("imageid not found");
+            }
+            return res.status(200).json({ status: "success", data });
+        })
+        .catch((err) => {
+            return res.status(404).json({ status: "error", message: "imageid not found" });
+        })
     }
-    return res.status(200).json({ status: "success", data: image });
-}
 
 
 const uploadImage = async (req, res) => {
@@ -33,7 +37,12 @@ const uploadImage = async (req, res) => {
         const name = uuidv4();
         const ext = path.extname(file[key].name);
         const imageName = name + ext;
-        const fileUrl = path.join("files", imageName)
+        let fileUrl = path.join("files", imageName)
+
+        if (process.platform == "win32") { 
+            fileUrl = fileUrl.replace("\\", "/");
+        }
+
         const filePath = path.join(__dirname, "..", "files", imageName);
         const image = new ImageModel({ url: fileUrl});
         await image.save()
@@ -53,6 +62,10 @@ const deleteImageById = async(req, res) => {
 
     await ImageModel.findById(videoId)
         .then(async (data) => {
+            if (!data) {
+                return res.status(404).json({ status: "error", message: `${videoId} not found` });
+            }
+
             fs.unlink(path.join(__dirname, "..", data.url), (err) => {
                 if (err) {
                     throw new Error("Could not delete video from file storage");
@@ -62,8 +75,9 @@ const deleteImageById = async(req, res) => {
             return res.status(200).json({ status: "success", message: `${videoId} deleted` })
         })
         .catch((err) => {
+            console.log(err)
             return res.status(500).json({ status: "error", message: err });
-        })
+        });
     // return res.status(200).json({ message: "success", data: [] });
 }
